KOGITO-2298: Renamed components to meet standards
