--- packages/management-console/src/components/Molecules/DataToolbarComponent/__mocks__/DataToolbarComponent.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const DataToolbarComponent = (): React.ReactElement => {
-  return <></>
-};
-
-export default DataToolbarComponent;
