+++ packages/management-console/src/components/Molecules/ProcessListToolbar/__mocks__/ProcessListToolbar.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProcessListToolbar = (): React.ReactElement => {
+  return <></>
+};
+
+export default ProcessListToolbar;
