--- packages/management-console/src/components/Organisms/DataListComponent/__mocks__/DataListComponent.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const DataListComponent = (): React.ReactElement => {
-  return <></>
-};
-
-export default DataListComponent;
