+++ packages/management-console/src/components/Organisms/ProcessListTable/__mocks__/ProcessListTable.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProcessListTable = (): React.ReactElement => {
+  return <></>
+};
+
+export default ProcessListTable;
