--- packages/management-console/src/components/Templates/DataListContainer/__mocks__/DataListContainer.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const DataListContainer = (): React.ReactElement => {
-  return <></>
-};
-
-export default DataListContainer;
