+++ packages/management-console/src/components/Templates/ProcessListPage/__mocks__/ProcessListPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProcessListPage = (): React.ReactElement => {
+  return <></>
+};
+
+export default ProcessListPage;
